requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
httpx>=0.27,<0.28
python-dotenv>=1.0,<2.0
aiosqlite>=0.19,<0.20
apscheduler>=3.10,<3.11
sendgrid>=6.11,<6.12
langdetect==1.0.9
trafilatura>=1.9,<2.0
beautifulsoup4>=4.12,<5.0
lxml>=5.2,<6.0
tldextract>=5.1.2
